Added method to skip cleaning of output assemblies to ExtensionsMetadataGenerator.

+++ tools/ExtensionsMetadataGenerator/src/ExtensionsMetadataGenerator/BuildTasks/RemoveRuntimeDependencies.cs
@@ -21,8 +21,17 @@ namespace Microsoft.Azure.WebJobs.Script.ExtensionsMetadataGenerator.BuildTasks
         [Required]
         public string OutputPath { get; set; }
 
+        [Required]
+        public ITaskItem[] IgnoreFiles { get; set; }
+
         public override bool Execute()
         {
+            HashSet<string> ignoreFilesSet = new HashSet<string>();
+            foreach (ITaskItem item in IgnoreFiles)
+            {
+                ignoreFilesSet.Add(item.ItemSpec);
+            }
+
             Assembly assembly = typeof(RemoveRuntimeDependencies).Assembly;
             using (Stream resource = assembly.GetManifestResourceStream(assembly.GetName().Name + ".runtimeassemblies.txt"))
             using (var reader = new StreamReader(resource))
@@ -32,7 +41,7 @@ namespace Microsoft.Azure.WebJobs.Script.ExtensionsMetadataGenerator.BuildTasks
                 {
                     string fileName = Path.Combine(OutputPath, assemblyName);
 
-                    if (File.Exists(fileName))
+                    if (File.Exists(fileName) && !ignoreFilesSet.Contains(assemblyName))
                     {
                         File.Delete(fileName);
                     }
